first part of restructuring
I decided to do proper OO structuring for some parts of cbob, because
making changes really became invasive in a lot of places. That means
good-bye for an easy conversion to C, though C++ is still an option (but
unlikely, really - I'm starting to use too much of the Python goodies).

Mostly, there are two classes - a '_Project' one (the underscore means
the user is not supposed to create an object himself; access to a
singleton object is through 'get_project()'), and a 'Target' one
(accessed primarily through the 'process.targets' property). The
'pathhelpers' module is going to shrink significantly, maybe even go
the way of the Dodo. So is, by the way, the 'checks' module: We're doing
proper, though very shallow, Exception handling stuff for errors.

Also, the 'print()' function is only used when the user asks for
information (e.g. 'cbob list'). For every other occasion we are using
the 'logging' module now.